+++ xodus/src/xvd/mod.rs
@@ -1,8 +1,5 @@
 mod crypt;
 pub mod math;
 pub mod streaming;
-pub mod streaming2;
-pub mod streaming3;
-pub mod streaming4;
 pub mod streaming_ntfs;
 pub mod utils;
